# perf: using Node.js compile cache