--- a/packages/eslint-config-eslint/package.json
+++ b/packages/eslint-config-eslint/package.json
@@ -33,7 +33,7 @@
     "@eslint-community/eslint-plugin-eslint-comments": "^4.3.0",
     "@eslint/js": "^9.0.0",
     "eslint-plugin-jsdoc": "^48.2.3",
-    "eslint-plugin-n": "^17.2.0",
+    "eslint-plugin-n": "^17.11.1",
     "eslint-plugin-unicorn": "^52.0.0"
   },
   "keywords": [