it was introduced in node.js v22.8.0
https://nodejs.org/en/blog/release/v22.8.0#new-js-api-for-compile-cache

fixes https://github.com/eslint/eslint/issues/18968

<!--
    Thank you for contributing!

    ESLint adheres to the [OpenJS Foundation Code of Conduct](https://eslint.org/conduct).
-->

#### Prerequisites checklist

- [x] I have read the [contributing guidelines](https://github.com/eslint/eslint/blob/HEAD/CONTRIBUTING.md).

#### What is the purpose of this pull request? (put an "X" next to an item)

<!--
    The following template is intentionally not a markdown checkbox list for the reasons
    explained in https://github.com/eslint/eslint/pull/12848#issuecomment-580302888
-->

[ ] Documentation update
[ ] Bug fix ([template](https://raw.githubusercontent.com/eslint/eslint/HEAD/templates/bug-report.md))
[ ] New rule ([template](https://raw.githubusercontent.com/eslint/eslint/HEAD/templates/rule-proposal.md))
[ ] Changes an existing rule ([template](https://raw.githubusercontent.com/eslint/eslint/HEAD/templates/rule-change-proposal.md))
[ ] Add autofix to a rule
[ ] Add a CLI option
[ ] Add something to the core
[ x] Other, please explain:

<!--
    If the item you've checked above has a template, please paste the template questions below and answer them. (If this pull request is addressing an issue, you can just paste a link to the issue here instead.)
-->

<!--
    Please ensure your pull request is ready:

    - Read the pull request guide (https://eslint.org/docs/latest/contribute/pull-requests)
    - Include tests for this change
    - Update documentation for this change (if appropriate)
-->

<!--
    The following is required for all pull requests:
-->

#### What changes did you make? (Give an overview)

tested on
* ESLint: v9.12.0
* Node version: v22.9.0
* Operating System: darwin 24.1.0

before the change: 
```bash
Loading:
  Load performance Run #1:  181.045881ms
  Load performance Run #2:  188.315573ms
  Load performance Run #3:  411.307467ms
  Load performance Run #4:  208.031565ms
  Load performance Run #5:  240.104659ms

  Load Performance median:  208.031565ms
```
after the change(2rd time): 
```bash
Loading:
  Load performance Run #1:  123.129709ms
  Load performance Run #2:  136.558748ms
  Load performance Run #3:  134.139566ms
  Load performance Run #4:  136.995985ms
  Load performance Run #5:  206.45432ms

  Load Performance median:  136.558748ms
```

#### Is there anything you'd like reviewers to focus on?
`module.enableCompileCache()` is still experimental [Stability: 1](https://nodejs.org/docs/latest/api/documentation.html#stability-index).1 - Active Development

https://nodejs.org/docs/latest/api/module.html#moduleenablecompilecachecachedir
<!-- markdownlint-disable-file MD004 -->