@mdjermanovic 2024-10-15
> `module.enableCompileCache()` is still experimental [Stability: 1](https://nodejs.org/docs/latest/api/documentation.html#stability-index).1 - Active Development

Then maybe we shouldn't use it yet? Description for Stability: 1 says "Use of the feature is not recommended in production environments".