@snitin315 2024-10-15
> > `module.enableCompileCache()` is still experimental [Stability: 1](https://nodejs.org/docs/latest/api/documentation.html#stability-index).1 - Active Development
> 
> Then maybe we shouldn't use it yet? Description for Stability: 1 says "Use of the feature is not recommended in production environments".

I think it should be fine to use, especially since [TypeScript recently implemented](https://github.com/microsoft/TypeScript/pull/59720) this too. We can always keep an eye out for any potential changes or improvements.