@mdjermanovic 2024-10-15
> I think it should be fine to use, especially since [TypeScript recently implemented](https://github.com/microsoft/TypeScript/pull/59720) this too. We can always keep an eye out for any potential changes or improvements.

Looks like they wrapped the `enableCompileCache()` call in a try-catch?

https://github.com/microsoft/TypeScript/blob/40caf3431986884b3da7a607de22ced5d364d9fc/Herebyfile.mjs#L332-L337

And there's also a config option to disable it:

https://github.com/microsoft/TypeScript/blob/40caf3431986884b3da7a607de22ced5d364d9fc/Herebyfile.mjs#L320