@aladdin-add 2024-10-16
It depends on what the result of a failed call is - if it's that in some cases the cache won't take effect without throwing an error - I think that's perfectly acceptable!