@mdjermanovic 2024-10-17
> It depends on what the result of a failed call is - if it's that in some cases the cache won't take effect, and _Not_ throwing an error - I think that's perfectly acceptable!

If the call ends up having no effect, I also think that's perfectly acceptable. I'm only concerned if module loading will start throwing errors or malfunctioning in any other way at some later point, as this is an experimental feature.