@aladdin-add 2024-10-17
its docs said:

> Since compile cache is supposed to be a quiet optimization that is not required for the application to be functional, this method is designed to not throw any exception when the compile cache cannot be enabled.

https://nodejs.org/docs/latest/api/module.html#moduleenablecompilecachecachedir

Just in case having an issue,  it's no harm to put it in a `try...catch`. wdyt?