@mdjermanovic 2024-10-17
> its docs said:
> 
> > Since compile cache is supposed to be a quiet optimization that is not required for the application to be functional, this method is designed to not throw any exception when the compile cache cannot be enabled.
> 
> https://nodejs.org/docs/latest/api/module.html#moduleenablecompilecachecachedir

I think the statement refers to the API: `enableCompileCache()` call will not throw an error when the compile cache cannot be enabled. But since it's still marked as an experimental feature, I'm not sure if it's guaranteed that everything will work properly when the compile cache is enabled.