@mdjermanovic 2024-10-17
> Hmm, can we also implement this with a flag, something like `eslint --flag enable-compile-cache`?

That might be a good option, though I think users can already manually enable this by setting the `NODE_COMPILE_CACHE` environment variable.