@aladdin-add 2024-10-17
well, now we have 3 choices：

1. enable it by default, and keep an eye out for any potential changes or improvements..
2. wait a while for it to stabilize before enabling it.
3. enable it when specifying ` --flag enable-compile-cache` (not enabled by default)

my preference is 2 > 1 > 3.