@nzakas 2024-10-17
The API is the part that's experimental, the feature has been around for a while but only enabled via environment variable. So the compile cache itself isn't experimental.